# Conflicts:
#	.gitignore

+++ RadElement.Service/ElementSetService.cs
@@ -2,13 +2,12 @@
 using RadElement.Core.Domain;
 using RadElement.Core.DTO;
 using RadElement.Core.Services;
-using Microsoft.EntityFrameworkCore;
+using Serilog;
+using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
-using Serilog;
 using System.Net;
-using System;
+using System.Threading.Tasks;
 
 namespace RadElement.Service
 {
@@ -36,13 +35,13 @@ namespace RadElement.Service
         {
             try
             {
-                var sets = await radElementDbContext.ElementSet.ToListAsync();
-                return new JsonResult(sets, HttpStatusCode.OK);
+                var sets = radElementDbContext.ElementSet.ToList();
+                return await Task.FromResult(new JsonResult(GetElementSetDetailsArrayDto(sets), HttpStatusCode.OK));
             }
             catch (Exception ex)
             {
                 logger.Error(ex, "Exception in method 'GetSets()'");
-                return new JsonResult(ex, HttpStatusCode.InternalServerError);
+                return await Task.FromResult(new JsonResult(ex, HttpStatusCode.InternalServerError));
             }
         }
 
@@ -51,26 +50,27 @@ namespace RadElement.Service
         /// </summary>
         /// <param name="setId">The set identifier.</param>
         /// <returns></returns>
-        public async Task<JsonResult> GetSet(int setId)
+        public async Task<JsonResult> GetSet(string setId)
         {
             try
             {
-                var sets = await radElementDbContext.ElementSet.ToListAsync();
-                var set = sets.Find(x => x.Id == setId);
+                if (IsValidSetId(setId))
+                {
+                    int id = Convert.ToInt32(setId.Remove(0, 4));
+                    var sets = radElementDbContext.ElementSet.ToList();
+                    var set = sets.Find(x => x.Id == id);
 
-                if (set != null)
-                {
-                    return new JsonResult(set, HttpStatusCode.OK);
+                    if (set != null)
+                    {
+                        return await Task.FromResult(new JsonResult(set, HttpStatusCode.OK));
+                    }
                 }
-                else
-                {
-                    return new JsonResult(string.Format("No such set with id '{0}'", setId), HttpStatusCode.NotFound);
-                }
+                return await Task.FromResult(new JsonResult(string.Format("No such set with id '{0}'.", setId), HttpStatusCode.NotFound));
             }
             catch (Exception ex)
             {
                 logger.Error(ex, "Exception in method 'GetSet(int setId)'");
-                return new JsonResult(ex, HttpStatusCode.InternalServerError);
+                return await Task.FromResult(new JsonResult(ex, HttpStatusCode.InternalServerError));
             }
         }
 
@@ -85,27 +85,27 @@ namespace RadElement.Service
             {
                 if (!string.IsNullOrEmpty(searchKeyword))
                 {
-                    var sets = await radElementDbContext.ElementSet.ToListAsync();
+                    var sets = radElementDbContext.ElementSet.ToList();
                     var filteredSets = sets.FindAll(x => x.Name.ToLower().Contains(searchKeyword.ToLower()) || x.Description.ToLower().Contains(searchKeyword.ToLower()) ||
                                                 x.ContactName.ToLower().Contains(searchKeyword.ToLower())); ;
                     if (filteredSets != null && filteredSets.Any())
                     {
-                        return new JsonResult(filteredSets, HttpStatusCode.OK);
+                        return await Task.FromResult(new JsonResult(GetElementSetDetailsArrayDto(filteredSets), HttpStatusCode.OK));
                     }
                     else
                     {
-                        return new JsonResult(string.Format("No such set with keyword '{0}'.", searchKeyword), HttpStatusCode.NotFound);
+                        return await Task.FromResult(new JsonResult(string.Format("No such set with keyword '{0}'.", searchKeyword), HttpStatusCode.NotFound));
                     }
                 }
                 else
                 {
-                    return new JsonResult(string.Format("Keyword '{0}' given is invalid", searchKeyword), HttpStatusCode.BadRequest);
+                    return await Task.FromResult(new JsonResult(string.Format("Keyword '{0}' given is invalid", searchKeyword), HttpStatusCode.BadRequest));
                 }
             }
             catch (Exception ex)
             {
                 logger.Error(ex, "Exception in method 'SearchSet(string searchKeyword)'");
-                return new JsonResult(ex, HttpStatusCode.InternalServerError);
+                return await Task.FromResult(new JsonResult(ex, HttpStatusCode.InternalServerError));
             }
         }
 
@@ -120,7 +120,7 @@ namespace RadElement.Service
             {
                 if (content == null || string.IsNullOrEmpty(content.ModuleName) || string.IsNullOrEmpty(content.ContactName) || string.IsNullOrEmpty(content.Description))
                 {
-                    return new JsonResult("Element set is invalid", HttpStatusCode.BadRequest);
+                    return await Task.FromResult(new JsonResult("Element set is invalid", HttpStatusCode.BadRequest));
                 }
 
                 ElementSet set = new ElementSet()
@@ -128,24 +128,17 @@ namespace RadElement.Service
                     Name = content.ModuleName.Replace("_", " "),
                     Description = content.Description,
                     ContactName = content.ContactName,
+                    Status = "Proposed"
                 };
 
-                await radElementDbContext.ElementSet.AddAsync(set);
+                radElementDbContext.ElementSet.Add(set);
                 radElementDbContext.SaveChanges();
-
-                if (set.Id != 0)
-                {
-                    return new JsonResult(new SetIdDetails() { SetId = set.Id.ToString() }, HttpStatusCode.Created);
-                }
-                else
-                {
-                    return new JsonResult(new SetIdDetails() { SetId = set.Id.ToString() }, HttpStatusCode.BadRequest);
-                }
+                return await Task.FromResult(new JsonResult(new SetIdDetails() { SetId = "RDES" + set.Id.ToString() }, HttpStatusCode.Created));
             }
             catch (Exception ex)
             {
                 logger.Error(ex, "Exception in method 'CreateSet(CreateUpdateSet content)'");
-                return new JsonResult(ex, HttpStatusCode.InternalServerError);
+                return await Task.FromResult(new JsonResult(ex, HttpStatusCode.InternalServerError));
             }
         }
 
@@ -155,33 +148,38 @@ namespace RadElement.Service
         /// <param name="setId">The set identifier.</param>
         /// <param name="content">The content.</param>
         /// <returns></returns>
-        public async Task<JsonResult> UpdateSet(int setId, CreateUpdateSet content)
+        public async Task<JsonResult> UpdateSet(string setId, CreateUpdateSet content)
         {
             try
             {
-                if (content == null || string.IsNullOrEmpty(content.ModuleName) || string.IsNullOrEmpty(content.ContactName) || string.IsNullOrEmpty(content.Description))
+                if (IsValidSetId(setId))
                 {
-                    return new JsonResult("Element set is invalid", HttpStatusCode.BadRequest);
+                    int id = Convert.ToInt32(setId.Remove(0, 4));
+
+                    if (content == null || string.IsNullOrEmpty(content.ModuleName) || string.IsNullOrEmpty(content.ContactName) || string.IsNullOrEmpty(content.Description))
+                    {
+                        return new JsonResult("Element set is invalid", HttpStatusCode.BadRequest);
+                    }
+
+                    var elementSets = radElementDbContext.ElementSet.ToList();
+                    var elementSet = elementSets.Find(x => x.Id == id);
+
+                    if (elementSet != null)
+                    {
+                        elementSet.Name = content.ModuleName.Replace("_", " ");
+                        elementSet.Description = content.Description;
+                        elementSet.ContactName = content.ContactName;
+                        radElementDbContext.SaveChanges();
+                        return await Task.FromResult(new JsonResult(string.Format("Set with id {0} is updated.", setId), HttpStatusCode.OK));
+                    }
                 }
 
-                var elementSets = await radElementDbContext.ElementSet.ToListAsync();
-                var elementSet = elementSets.Find(x => x.Id == setId);
-
-                if (elementSet != null)
-                {
-                    elementSet.Name = content.ModuleName.Replace("_", " ");
-                    elementSet.Description = content.Description;
-                    elementSet.ContactName = content.ContactName;
-                    radElementDbContext.SaveChanges();
-                    return new JsonResult(string.Format("Set with id {0} is updated.", setId), HttpStatusCode.OK);
-                }
-
-                return new JsonResult(string.Format("No such set with id {0}.", setId), HttpStatusCode.NotFound);
+                return await Task.FromResult(new JsonResult(string.Format("No such set with id '{0}'.", setId), HttpStatusCode.NotFound));
             }
             catch (Exception ex)
             {
                 logger.Error(ex, "Exception in method 'CreateSet(CreateUpdaUpdateSet(int setId, CreateUpdateSet content)'");
-                return new JsonResult(ex, HttpStatusCode.InternalServerError);
+                return await Task.FromResult(new JsonResult(ex, HttpStatusCode.InternalServerError));
             }
         }
 
@@ -190,49 +188,88 @@ namespace RadElement.Service
         /// </summary>
         /// <param name="setId">The set identifier.</param>
         /// <returns></returns>
-        public async Task<JsonResult> DeleteSet(int setId)
+        public async Task<JsonResult> DeleteSet(string setId)
         {
             try
             {
-                var elementSets = await radElementDbContext.ElementSet.ToListAsync();
-                var elementSet = elementSets.Find(x => x.Id == setId);
-
-                if (elementSet != null)
+                if (IsValidSetId(setId))
                 {
-                    var elementSetRefs = radElementDbContext.ElementSetRef.ToList().FindAll(x => x.ElementSetId == elementSet.Id);
-                    if (elementSetRefs != null && elementSetRefs.Any())
+                    int id = Convert.ToInt32(setId.Remove(0, 4));
+                    var elementSets = radElementDbContext.ElementSet.ToList();
+                    var elementSet = elementSets.Find(x => x.Id == id);
+
+                    if (elementSet != null)
                     {
-                        foreach (var setref in elementSetRefs)
+                        var elementSetRefs = radElementDbContext.ElementSetRef.ToList().FindAll(x => x.ElementSetId == elementSet.Id);
+                        if (elementSetRefs != null && elementSetRefs.Any())
                         {
-                            var elementValues = radElementDbContext.ElementValue.ToList().FindAll(x => x.ElementId == setref.ElementId);
-                            var elements = radElementDbContext.Element.ToList().FindAll(x => x.Id == setref.ElementId);
-
-                            if (elementValues != null && elementValues.Any())
+                            foreach (var setref in elementSetRefs)
                             {
-                                radElementDbContext.ElementValue.RemoveRange(elementValues);
+                                var elementValues = radElementDbContext.ElementValue.ToList().FindAll(x => x.ElementId == setref.ElementId);
+                                var elements = radElementDbContext.Element.ToList().FindAll(x => x.Id == setref.ElementId);
+
+                                if (elementValues != null && elementValues.Any())
+                                {
+                                    radElementDbContext.ElementValue.RemoveRange(elementValues);
+                                }
+
+                                if (elements != null && elements.Any())
+                                {
+                                    radElementDbContext.Element.RemoveRange(elements);
+                                }
                             }
 
-                            if (elements != null && elements.Any())
-                            {
-                                radElementDbContext.Element.RemoveRange(elements);
-                            }
+                            radElementDbContext.ElementSetRef.RemoveRange(elementSetRefs);
                         }
 
-                        radElementDbContext.ElementSetRef.RemoveRange(elementSetRefs);
+                        radElementDbContext.ElementSet.Remove(elementSet);
+                        radElementDbContext.SaveChanges();
+                        return await Task.FromResult(new JsonResult(string.Format("Set with id {0} is deleted.", setId), HttpStatusCode.OK));
                     }
-
-                    radElementDbContext.ElementSet.Remove(elementSet);
-                    radElementDbContext.SaveChanges();
-                    return new JsonResult(string.Format("Set with id {0} is deleted.", setId), HttpStatusCode.OK);
                 }
-
-                return new JsonResult(string.Format("No such set with id {0}.", setId), HttpStatusCode.NotFound);
+                return await Task.FromResult(new JsonResult(string.Format("No such set with id '{0}'.", setId), HttpStatusCode.NotFound));
             }
             catch (Exception ex)
             {
                 logger.Error(ex, "Exception in method 'DeleteSet(int setId)'");
-                return new JsonResult(ex, HttpStatusCode.InternalServerError);
+                return await Task.FromResult(new JsonResult(ex, HttpStatusCode.InternalServerError));
             }
+        }
+
+        private bool IsValidSetId(string setId)
+        {
+            if (setId.Length > 4 && setId.Substring(0, 4) == "RDES")
+            {
+                int id;
+                bool result = Int32.TryParse(setId.Remove(0, 4), out id);
+                return result;
+            }
+
+            return false;
+        }
+
+        private List<ElementSetDetails> GetElementSetDetailsArrayDto(List<ElementSet> sets)
+        {
+            List<ElementSetDetails> setDetails = new List<ElementSetDetails>();
+            foreach (var set in sets)
+            {
+                setDetails.Add(GetElementSetDetailsDto(set));
+            }
+
+            return setDetails;
+        }
+
+        private ElementSetDetails GetElementSetDetailsDto(ElementSet set)
+        {
+            return new ElementSetDetails()
+            {
+                Id = "RDES" + set.Id,
+                ContactName = set.ContactName,
+                Description = set.Description,
+                Name = set.Name,
+                ParentId = set.ParentId,
+                Status = set.Status
+            };
         }
     }
 }
